Documented code sections
Added documentation in comments regarding the purpose of the project and
namespace.

Added comments to delimit each section of the namespace.  This makes it easier
to get up to speed with the namespace and make changes with confidence.
Distinguishing sections also makes it easier to know when to refactor sections
into their own namespace (and an idea of what to call that namespace).