scoreboard and score-card components added
* scoreboard
a component to manage the display of the overall scoreboard

* `score-card`
display details about the specific score using a bluma css box

* `scoreboard-source`
Defined a temporary data structure as a place holder for the actual data to be
received from the `game-scoreboard-api`.  Enables a visual test of the
`scoreboard` and `score-card` components.